Renamed folders, files, and classes
