+++ src/components/Filters/Filters.scss
@@ -1,6 +1,6 @@
 @import "../../scss/base";
 
-.filter {
+.filters {
   margin-top: 60px;
   margin-bottom: 60px;
   margin-left: 60px;
