+++ src/components/ItemsToDisplay/ItemsToDisplay.css
@@ -36,14 +36,14 @@ a {
   width: 24px;
 }
 
-.filter {
+.filters {
   margin-top: 60px;
   margin-bottom: 60px;
   margin-left: 60px;
   display: flex;
   gap: 24px;
 }
-.filter span {
+.filters span {
   border: 1px solid #282C2F;
   cursor: pointer;
 }
